requirements.txt
numpy
pandas
sqlalchemy
google-cloud-bigquery
db-types
xmltodict
dvc[gdrive]
easyocr
flask